added labManagementProject
An SQL, Python and Tableau Public project that has management of lab info. The python script creates fake data to insert into the database to use in Tableau. A Tableau graph is in the readme section and it is used to show which lab managers used how my pto compared to hours worked. Might update later to add fully defined functions to use.